started css for the member page, reset password, forgot password, and update info pages. added confirms for deleting dates and members. members will also show up in alphebetical order on the member page now

+++ client/src/pages/Dashboard/Dashboard.js
@@ -93,13 +93,15 @@ class Dashboard extends Component {
 
 	// deletes date
 	deleteDate = (id) => {
-		let dateID = {
-			id: id
-		}
-		API.deleteDate(dateID)
-		.then(res => this.loadDates())
-	    .catch(err => console.log(err));
-	    this.loadDates();
+		if(window.confirm('Are you sure you want to delete this date? All data will be deleted.')) {
+			let dateID = {
+				id: id
+			}
+			API.deleteDate(dateID)
+			.then(res => this.loadDates())
+		    .catch(err => console.log(err));
+		    this.loadDates();
+	    }
 	};
 
 	// handles form input
